# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Authentication API
  version: 1.0.0
  description: API documentation for user authentication system

servers:
  - url: http://localhost:3000
    description: Local development server

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token

  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - name
        - mbNum
        - password
        - passwordVerify
        - login_type
      properties:
        email:
          type: string
          example: user@example.com
        name:
          type: string
          example: John Doe
        mbNum:
          type: string
          example: "9876543210"
        password:
          type: string
          example: mypassword123
        passwordVerify:
          type: string
          example: mypassword123
        login_type:
          type: string
          example: customer

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: mypassword123

    ResetPasswordRequest:
      type: object
      required:
        - resetemail
        - oldPassword
        - newPassword
      properties:
        resetemail:
          type: string
          example: user@example.com
        oldPassword:
          type: string
          example: oldpassword123
        newPassword:
          type: string
          example: newpassword123

    ForgotPasswordRequest:
      type: object
      required:
        - forgotemail
        - forgotmobile
      properties:
        forgotemail:
          type: string
          example: user@example.com
        forgotmobile:
          type: string
          example: "9876543210"

paths:
  /:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: User registered successfully
        '400':
          description: Missing fields or validation error
        '500':
          description: Server error

  /login:
    post:
      summary: Login user and set JWT cookie
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful — token stored in cookie
        '400':
          description: Missing required fields
        '401':
          description: Invalid email or password
        '500':
          description: Server error

  /logout:
    get:
      summary: Logout user by clearing token cookie
      tags: [Auth]
      responses:
        '200':
          description: Logged out successfully

  /loggedIn:
    get:
      summary: Check if user is logged in
      tags: [Auth]
      responses:
        '200':
          description: Returns true or false
          content:
            application/json:
              schema:
                type: boolean

  /logintype:
    get:
      summary: Get authenticated user's login type
      tags: [Auth]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Returns user information
        '401':
          description: Unauthorized
        '500':
          description: Server error

  /resetpassword:
    post:
      summary: Reset password for logged-in user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password updated successfully
        '400':
          description: Missing required fields
        '401':
          description: Incorrect email or old password
        '500':
          description: Server error

  /forgotpassword:
    post:
      summary: Forgot password — send current password to email
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Mail sent successfully
        '400':
          description: Missing required fields
        '401':
          description: Email or phone number incorrect
        '500':
          description: Server error